requirements.txt
Django==4.2.5
django-cleanup==8.0.0
django-filter==23.2
djangorestframework==3.14.0
gunicorn==21.2.0
Pillow==10.0.1
psycopg2-binary==2.9.7

# testing
pytest==7.4.2
pytest-django==4.5.2
pytest-drf==1.1.3
pytest-schema==0.1.1

